Create directories for diagnostics and completions
